move openstack values to secret
